+++ src/Accordion/stories.js
@@ -1,8 +1,8 @@
 import React from 'react'
 
 import create from '../create'
-
 import Accordion from './'
+import Tag from '../Tag'
 
 create('Accordion')
   .add('Default', () => (
@@ -12,9 +12,21 @@ create('Accordion')
       </Accordion>
     </figure>
   ))
+  .add('Counter', () => (
+    <figure>
+      <Accordion title={
+        <>
+            This is the title
+            <Tag>23</Tag>
+        </>
+      }>
+        <p>Lorem ipsum dolor sit amet</p>
+      </Accordion>
+    </figure>
+  ))
   .add('Default open', () => (
     <figure>
-      <Accordion title="See more" open={true}>
+      <Accordion title="Opened" open={true}>
         <p>Lorem ipsum dolor sit amet</p>
       </Accordion>
     </figure>